added current thinkings to seeder

+++ database/migrations/2016_03_09_222301_add_new_currents_to_project.php
@@ -48,11 +48,12 @@ class AddNewCurrentsToProject extends Migration
         
         foreach ($categories as $category)
         {            
-            
-            Schema::table('ideas', function($table) use ($category)
-            {
-                $table->dropColumn($category->name);
-            });
+            if (Schema::hasColumn('ideas', $category->name))
+                Schema::table('ideas', function($table) use ($category)
+                {
+
+                    $table->dropColumn($category->name);
+                });
         }
         
         Schema::table('ideas', function($table)
